+++ gammapy/maps/hpx/tests/test_ndmap.py
@@ -9,7 +9,7 @@ from regions import CircleSkyRegion
 from gammapy.maps import HpxGeom, HpxMap, HpxNDMap, Map, MapAxis, WcsGeom
 from gammapy.utils.testing import mpl_plot_check, requires_data, requires_dependency
 from gammapy.maps.utils import find_bintable_hdu
-from gammapy.maps.hpx import HpxConv
+from gammapy.maps.hpx.io import HpxConv
 from gammapy.irf import PSFKernel, PSFMap
 
 pytest.importorskip("healpy")
